feat: Serializer scoped into mutils
