+++ lib/generators/mutils/USAGE
@@ -2,7 +2,7 @@ Description:
     Generates a serializer for the given model.
 
 Example:
-    rails generate serializer User
+    rails generate mutils:serializer User
 
     This will create:
         app/serializers/user_serializer.rb
