+++ lib/generators/mutils/templates/serializer.rb.tt
